# openapi.yaml
openapi: 3.0.1
info:
  title: Authentication API and Storage Files Manager API
  description: MSL Group production
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Generated server url
tags:
  - name: Files manager
    description: Endpoints for file management operations
  - name: Authentication
    description: Endpoints for user authentication and logout
paths:
  /file:
    get:
      tags:
        - Files manager
      summary: Downloads a file
      operationId: serveFile
      parameters:
        - name: fileRequest
          in: query
          description: The file to download
          required: true
          schema:
            $ref: '#/components/schemas/filename'
      responses:
        '200':
          description: Reference to file that can be served over HTTP
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request format or missing parameters or input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      tags:
        - Files manager
      summary: Renames a file
      operationId: handleFileRename
      parameters:
        - name: filename
          in: query
          description: The new filename for the file
          required: true
          schema:
            type: string
          example: newFileName.txt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/filename'
        required: true
      responses:
        '200':
          description: The renamed file name
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: Invalid request format or missing parameters or input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
        - Files manager
      summary: Uploads a file
      operationId: handleFileUpload
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: The file to upload
              format: binary
        required: true
      responses:
        '200':
          description: The uploaded file name
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: Invalid request format or missing parameters or input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
        - Files manager
      summary: Deletes a file
      operationId: delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/filename'
      responses:
        '200':
          description: The deleted file name
          content:
            '*/*':
              schema:
                type: string
        '400':
          description: Invalid request format or missing parameters or input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate JWT token
      operationId: logout
      responses:
        '200':
          description: Successful logout
  /login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Returns a token for a valid user
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/credentials'
        required: true
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth-token'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /list:
    get:
      tags:
        - Files manager
      summary: Returns list of uploaded files with limit parameter
      operationId: limitListUploadedFiles
      parameters:
        - name: limit
          in: query
          description: Limit(min=1)number of files returned
          required: true
          schema:
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: List of file-info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-info'
        '400':
          description: Invalid request format or missing parameters or input data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  schemas:
    filename:
      required:
        - filename
      type: object
      properties:
        filename:
          type: string
      description: The file to delete
    error:
      type: object
      properties:
        message:
          type: string
        id:
          type: integer
          format: int64
      description: Represent an error that may occur when processing request
    credentials:
      required:
        - login
        - password
      type: object
      properties:
        password:
          maxLength: 64
          minLength: 6
          type: string
        login:
          maxLength: 64
          minLength: 6
          type: string
      description: Contains non-empty login validated by the @Email constraint
    auth-token:
      type: object
      properties:
        auth-token:
          type: string
      description: Contains json web token
    file-info:
      type: object
      properties:
        filename:
          type: string
        size:
          type: integer
          format: int64
      description: Contains file information